Autoconfiscate Bulwark - Part 1
Not currently working but as I migrate to the Autotools, I've already spent a lot of time rearranging files and the like. I've also made early versions of the *.am files and configure.ac so making a commit now just to try and break up the changes.

The new structure is to have all public API header files in include/, with all other sources and headers in src/. Test code is in tests/ and example code is in examples/. The bootstrap.sh shell script is the entry point to building the project.