Using poll syscall instead of pthread
Found that using pthread was overkill, can instead use poll to see if
input is ready to be read, and read if so.

Calling Bulwark_PollEvents fills bulwarks event queue if any input
events happened, which you can then read in a loop with
Bulwark_ReadNextEventInQueue(output);

Also fixed a logging bug that was hard to find, needed to use vfprint
since I was passing variables arguments to the printf.

Fixed a bug where calling Bulwark_SetCursorVisible always did the
opposite of what you would expect.